refactor(agent): fold fragments, thread real ids into events
- core/policy/engine.ts (pass-through over GenericPolicyEngine) -> policy/index.ts
- core/prompt-builder.ts -> execution/turn-prepare.ts (its single caller)
- core/execution/shared.ts dispersed: resolveProviderModel -> runner.ts,
  toMessagesWithParts -> run-state.ts, summarizeInput -> tool-executor.ts
  (tool-guard repointed)
- core/chat-agent.ts KEPT: package public entry; folding would push
  streamAgent internals into the barrel
- replace hardcoded "chat-agent" ids: budget/compaction policies emit the
  dispatch-context sessionId (run.turn.pre / run.completion.pre contracts
  guarantee it); compaction summary message carries the source message's
  agent id; pinned by a red-first test on the budget warning event
