--- packages/agent/src/core/prompt-builder.ts
@@ -1,15 +1,0 @@
-import type { Tool } from "@openomni/protocol";
-
-export function buildSystemPrompt(
-  basePrompt: string | undefined,
-  tools: Tool.Spec[],
-): string | undefined {
-  const toolPrompts = tools
-    .filter((t) => t.prompt)
-    .map((t) => `## Tool: ${t.name}\n${t.prompt}`)
-    .join("\n\n");
-
-  if (!toolPrompts) return basePrompt;
-  if (!basePrompt) return toolPrompts;
-  return `${basePrompt}\n\n---\n\n${toolPrompts}`;
-}
